Use pointers to improve stability in TrackingNode
Earlier we had issues upon deletion of the threads/osc ports.
This change increases the stability by deleting the objects correctly
and makes sure we don't try to open the ports immediately, allowing the
old objects to be deleted first when opening a configuration.
This is because OpenEphys adds new things before deleting the old.